Ajusta retornos dos erros

+++ services/ProdutoService.js
@@ -1,6 +1,7 @@
 const { PrismaClient } = require("@prisma/client");
 const StringUtils = require("../utils/StringUtils");
 const utils = new StringUtils();
+const ProdutoNaoEncontradoError = require("../utils/ProdutoNaoEncontradoError")
 
 class ProdutoService {
   constructor() {
@@ -8,21 +9,29 @@ class ProdutoService {
   }
 
   async listar() {
-    const produtos = await this.prisma.produto.findMany();
-    return produtos;
+    try {
+      const produtos = await this.prisma.produto.findMany();
+      return produtos;
+    } catch (error) {
+      throw new Error('Erro ao listar produtos: ' + error.message);
+    }
   }
 
   async listarPorId(id) {
     const produto = await this.prisma.produto.findFirst({
       where: { id },
     });
+
+    if (!produto) {
+      throw new ProdutoNaoEncontradoError();
+    }
+
     return produto;
   }
 
   async cadastrar(produto) {
     const nomeNormalizado = utils.removerAcentos(produto.nome).toLowerCase();
-    
-    // Validação para evitar produtos duplicados
+
     const produtoExistente = await this.prisma.produto.findFirst({
       where: {
         AND: [
@@ -33,21 +42,17 @@ class ProdutoService {
     });
 
     if (produtoExistente) {
-      throw new Error('Já existe um produto com este nome nesta categoria');  
+      throw new Error('Já existe um produto com este nome nesta categoria');
     }
-
-    // Verificar se a categoria existe
 
     const categoria = await this.prisma.categoria.findFirst({
       where: { id: produto.categoria_id },
     });
 
-  
+
     if (!categoria) {
       throw new Error('Categoria não encontrada.');
     }
-
-    // Criar o produto
 
     const novoProduto = await this.prisma.produto.create({
       data: {
@@ -56,8 +61,6 @@ class ProdutoService {
         categoria_id: produto.categoria_id,
       },
     });
-
-    // Atualizar a categoria
 
     await this.prisma.categoria.update({
       where: { id: produto.categoria_id },
@@ -69,46 +72,78 @@ class ProdutoService {
         },
       },
     });
+
+    return novoProduto;
   }
 
   async editar(id, produto) {
-    const nomeNormalizado = utils.removerAcentos(produto.nome).toLowerCase();
-    
-    // Validação para evitar produtos duplicados
-    
-    const produtoExistente = await this.prisma.produto.findFirst({
-      where: {
-        AND: [
-          { nome: { equals: nomeNormalizado } },
-          { categoria_id: { equals: produto.categoria_id } },
-        ],
-      },
-    });
+    try {
+      const nomeNormalizado = utils.removerAcentos(produto.nome).toLowerCase();
 
-    if (produtoExistente) {
-      throw new Error('Já existe um produto com este nome nesta categoria');  
+      const produtoExistente = await this.prisma.produto.findFirst({
+        where: {
+          AND: [
+            { nome: { equals: nomeNormalizado } },
+            { categoria_id: { equals: produto.categoria_id } },
+          ],
+        },
+      });
+
+      if (produtoExistente) {
+        throw new Error('Já existe um produto com este nome nesta categoria');
+      }
+
+      const categoria = await this.prisma.categoria.findFirst({
+        where: { id: produto.categoria_id },
+      });
+
+      if (!categoria) {
+        throw new Error('Categoria não encontrada.');
+      }
+
+      const produtoEncontrado = await this.prisma.produto.findUnique({
+        where: { id },
+      });
+
+      if (!produtoEncontrado) {
+        throw new Error('Produto com o ID ' + id + ' não encontrado.');
+      }
+
+      const produtoAtualizado = await this.prisma.produto.update({
+        where: { id },
+        data: {
+          nome: produto.nome.toLowerCase(),
+          descricao: produto.descricao,
+          categoria_id: produto.categoria_id,
+        },
+      });
+
+      return produtoAtualizado;
+    } catch (error) {
+      throw new Error(error.message);
     }
-
-    const categoria = await this.prisma.categoria.findFirst({
-      where: { id: produto.categoria_id },
-    });
-
-    if (!categoria) {
-      throw new Error('Categoria não encontrada.');
-    }
-
-    await this.prisma.produto.update({
-      where: { id },
-      data: {
-        nome: produto.nome.toLowerCase(),
-        descricao: produto.descricao,
-        categoria_id: produto.categoria_id,
-      },
-    });
   }
 
   async excluir(id) {
-    await this.prisma.produto.delete({ where: { id } });
+    try {
+      const produto = await this.prisma.produto.findFirst({
+        where: { id },
+      });
+
+      if (!produto) {
+        throw new ProdutoNaoEncontradoError();
+      }
+
+      await this.prisma.produto.delete({ where: { id } });
+
+      return { message: 'Produto excluído com sucesso.' };
+    } catch (error) {
+      if (error.code === 'P2003') {
+        throw new Error('Erro ao excluir produto: Produto não encontrado.');
+      } else {
+        throw new Error('Erro ao excluir produto: ' + error.message);
+      }
+    }
   }
 }
